+++ README.md
@@ -1,13 +1,13 @@
-[![CI](https://github.com/cassidymwagner/oceans_sf/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/cassidymwagner/oceans_sf/actions/workflows/ci.yml)
+[![CI](https://github.com/cassidymwagner/FluidSF/actions/workflows/ci.yml/badge.svg?branch=main)](https://github.com/cassidymwagner/FluidSF/actions/workflows/ci.yml)
 [![DOI](https://zenodo.org/badge/DOI/10.5281/zenodo.10463864.svg)](https://doi.org/10.5281/zenodo.10463864)
 
 
-# Oceans-SF
- Oceans-SF is a package used to calculate structure functions from oceanographic data. 
+# FluidSF
+ FluidSF is a package used to calculate structure functions from oceanographic data. 
 
 Installation
 ---
-Fork/clone this repository to your local machine. The easiest method to install Oceans-SF is with [pip](https://pip.pypa.io/):
+Fork/clone this repository to your local machine. The easiest method to install FluidSF is with [pip](https://pip.pypa.io/):
 
 ```console
 $ pip install .
